+++ src/core/hle/kernel/config_mem.cpp
@@ -3,7 +3,7 @@
 // Refer to the license.txt file included.
 
 #include <cstring>
-#include "core/hle/config_mem.h"
+#include "core/hle/kernel/config_mem.h"
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
