+++ src/core/hle/kernel/config_mem.h
