+++ src/core/hle/kernel/memory.cpp
@@ -12,7 +12,7 @@
 #include "common/common_types.h"
 #include "common/logging/log.h"
 #include "core/core.h"
-#include "core/hle/config_mem.h"
+#include "core/hle/kernel/config_mem.h"
 #include "core/hle/kernel/memory.h"
 #include "core/hle/kernel/vm_manager.h"
 #include "core/hle/result.h"
